Fix the flux instantiation issue
Calling `flux_new` would fail in init and in the `CVDisplayLink`
callback. I think this was mostly related to not calling
`makeCurrentContext` before instantiating flux (because we do call
OpenGL APIs on init), and, perhaps, related to not setting `view` on
the `NSOpenGLContext` at the right time. Or maybe it was party a
threading issue, but I think I’ve figured out how to set up a
`CGDisplayLink` loop correctly. More or less.