Fix the framerate and enable vsync
This thread::sleep was a temporary fix, but wasn’t actually necessary.
You can just let the program run off and it will happily run at 500fps.
Because of it, though, the screensaver was running at 48fps, instead of
60.

The right way to framelimit with OpenGL is to set the swap interval and
enable vsync.