Fix rendering on Intel GPUs
The Intel GPU I have (a rather old Iris 5100) reports that it supports
NV_DX_interop2, but, no matter what I do, I can't get it to register the
DX color buffer.

I managed to register a plain texture, but it still fails when trying to
attach it to a framebuffer.